requirements-qwen.txt
# Qwen3-TTS Dependencies (Windows Local)
#
# Setup (requires Python 3.12, CUDA 12.8):
#   1. Install PyTorch with CUDA:
#      pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
#   2. Install dependencies:
#      pip install -r requirements.txt
#   3. (Optional) For faster inference, install flash-attn prebuilt wheel:
#      pip install https://github.com/kingbri1/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu128torch2.8.0cxx11abiFALSE-cp312-cp312-win_amd64.whl

transformers>=4.57.3
accelerate>=1.12.0
# bitsandbytes>=0.45.0
einops
gradio
librosa
torchaudio
soundfile
onnxruntime
numpy